Commit Part4 exercise 4.21 bloglist expansion step 9
Part4 exercise 4.21 bloglist expansion step 9

+++ part4/bloglistbackend/controllers/blogs.js
@@ -38,13 +38,24 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+  if ( blog.author.toString() !== user.id.toString() ) {
+    return response.status(401).json({ error: 'only the owner of the blog can delete it' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const blog = request.body
-
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
   response.json(updatedBlog)
 })
